fix: resolve 17 mypy type errors in CI lint job
- figure.py: add None guards for rcParams.get(), cast tuple() for
  figsize, int() for dpi, typed kwargs dict
- layout.py: type: ignore[assignment] for GridSpec cast
- __init__.py: type: ignore[method-assign] for Axes.twinx patch
- formatting.py: change fontsize param type from int to float
- validate_enhanced.py: add list[str] type annotation
- mcp/tools.py: type: ignore[arg-type] for to_hex tuple arg

+++ src/dartwork_mpl/helpers/formatting.py
@@ -119,7 +119,7 @@ def add_value_labels(
     format_str: str = ".1f",
     offset_y: float = 0.02,
     color: str | None = None,
-    fontsize: int | None = None,
+    fontsize: float | None = None,
 ) -> None:
     """Add value labels to data points.
